#215 : moving joda code from the oracle extension to the oracle-joda-time extension
