+++ extensions/oracle-joda-time/main/resources/META-INF/services/org.sql2o.converters.ConvertersProvider
@@ -8,4 +8,6 @@
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 
-org.sql2o.converters.OracleDateConverter
+org.sql2o.converters.joda.OracleDateTimeConverter
+org.sql2o.converters.joda.OracleLocalDateConverter
+org.sql2o.converters.joda.OracleLocalTimeConverter
