+++ extensions/oracle-joda-time/pom.xml
@@ -10,45 +10,32 @@
 
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.sql2o.extensions</groupId>
-    <artifactId>extensions-parent</artifactId>
-    <name>Sql2o - Extensions</name>
-    <packaging>pom</packaging>
-
+    <artifactId>sql2o-oracle-joda-time</artifactId>
+    <packaging>jar</packaging>
+    <name>sql2o-oracle-joda-time</name>
+    <description>Joda-time extension for Oracle for sql2o</description>
+    <url>http://www.sql2o.org</url>
     <parent>
-        <groupId>org.sql2o</groupId>
-        <artifactId>sql2o-parent</artifactId>
+        <groupId>org.sql2o.extensions</groupId>
+        <artifactId>extensions-parent</artifactId>
         <version>1.6.0-RC3-SNAPSHOT</version>
     </parent>
 
-    <modules>
-        <module>postgres</module>
-        <module>oracle</module>
-        <module>oracle-joda-time</module>
-        <module>db2</module>
-    </modules>
-
     <dependencies>
-        <!-- All dependencies depends on sql2o core project -->
         <dependency>
-            <groupId>org.sql2o</groupId>
-            <artifactId>sql2o</artifactId>
-            <version>${project.version}</version>
+            <groupId>com.oracle</groupId>
+            <artifactId>ojdbc6</artifactId>
+            <version>11.2.0.4</version>
         </dependency>
+
+        <!-- optional dependencies -->
+        <dependency>
+            <groupId>joda-time</groupId>
+            <artifactId>joda-time</artifactId>
+            <version>2.0</version>
+            <optional>true</optional>
+        </dependency>
+
     </dependencies>
 
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>2.4</version>
-                <configuration>
-                    <archive>
-                        <manifestFile>src/main/resources/META-INF/MANIFEST.MF</manifestFile>
-                    </archive>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
 </project>
